refactor: remove react-markdown
