+++ pages/lp/[slug].tsx
@@ -1,6 +1,5 @@
 import { GetStaticPaths, GetStaticProps } from 'next';
 import React, { ReactElement } from 'react';
-import { ReactMarkdown } from 'react-markdown/lib/react-markdown';
 import { diamond } from '#components-ui/logo/home-h1';
 import AdministrationDescription from '#components/administrations/administration-description';
 import { LayoutDefault } from '#components/layouts/layout-default';
@@ -12,6 +11,7 @@ import {
   getLandingPage,
 } from '#models/landing-pages';
 import { IPropsWithMetadata } from '#utils/server-side-props-helper/post-server-side-props';
+import parseMarkdownSync from '#utils/static-markdown-pages/parse-markdown';
 import { NextPageWithLayout } from 'pages/_app';
 
 interface IProps extends IPropsWithMetadata, ILandingPage {}
@@ -67,13 +67,21 @@ const LandingPage: NextPageWithLayout<IProps> = ({
           <div key={block.title} className="fr-col-12 fr-col-sm-4 fr-col-md-4">
             <div className="reassurance">
               <h2>{block.title}</h2>
-              <ReactMarkdown>{block.body}</ReactMarkdown>
+              <div
+                dangerouslySetInnerHTML={{
+                  __html: parseMarkdownSync(block.body).html,
+                }}
+              />{' '}
             </div>
           </div>
         ))}
       </div>
       <br />
-      {body && <ReactMarkdown>{body}</ReactMarkdown>}
+      <div
+        dangerouslySetInnerHTML={{
+          __html: parseMarkdownSync(body).html,
+        }}
+      />{' '}
       {datasources.length > 0 && (
         <h2>Quelles sont les sources des données utilisées ?</h2>
       )}
