+++ utils/static-markdown-pages/load-all.ts
@@ -1,11 +1,5 @@
-import remarkHeadings from '@vcarl/remark-headings';
-import rehypeStringify from 'rehype-stringify';
-// @ts-ignore
-import remarkHeadingId from 'remark-heading-id';
-import remarkParse from 'remark-parse';
-import remarkRehype from 'remark-rehype';
-import { unified } from 'unified';
 import { IArticle } from '#models/article/type';
+import parseMarkdownSync from './parse-markdown';
 
 export function loadAll<T extends IArticle>(
   articlesFolderContext: Record<string, T>
@@ -26,31 +20,6 @@ export function loadAll<T extends IArticle>(
 
   return rawArticles.map((article) => ({
     ...article,
-    body: parseBodyMarkdown(article.body as unknown as string),
+    body: parseMarkdownSync(article.body as unknown as string),
   }));
 }
-
-function parseBodyMarkdown(body: string): {
-  html: string;
-  raw: string;
-  headings: { id: string; value: string; depth: number }[];
-} {
-  const parsedBody = markdownProcessor.processSync(body);
-  return {
-    raw: body,
-    html: parsedBody.value as string,
-    // @ts-ignore
-    headings: parsedBody.data.headings.map((header) => ({
-      id: header.data.id,
-      value: header.value,
-      depth: header.depth,
-    })),
-  };
-}
-
-const markdownProcessor = unified()
-  .use(remarkParse)
-  .use(remarkHeadingId, { defaults: true })
-  .use(remarkHeadings)
-  .use(remarkRehype)
-  .use(rehypeStringify);
